requirements.txt
redis
requests
python-telegram-bot
logtail-python
logentries
python-decouple